replace example with poc for class Hitzone.java
