--- src/main/java/de/midorlo/relentless/domain/items/CustomerRepository.java
@@ -1,12 +1,0 @@
-package de.midorlo.relentless.domain.items;
-
-import java.util.List;
-
-import org.springframework.data.repository.CrudRepository;
-
-public interface CustomerRepository extends CrudRepository<Customer, Long> {
-
-    List<Customer> findByLastName(String lastName);
-
-    Customer findById(long id);
-}
